+++ app/src/main/java/com/example/lumen/domain/ble/model/LedControllerState.kt
@@ -5,7 +5,7 @@ package com.example.lumen.domain.ble.model
  */
 data class LedControllerState(
     val isOn: Boolean,
-    val preset: Byte,
+    val preset: Int,
     val speed: Byte,
     val brightness: Float,
     val icModel: IcModel,
